==> Notebooks/1) Data_Forming.ipynb <==
import pandas as pd
import numpy as np

#read raw files as csv files

electricity_data = pd.read_csv('Data/Raw data/electricity_cleaned.txt', sep=',')

metadata_data = pd.read_csv('Data/Raw data/metadata.txt', sep=',')

weather_data = pd.read_csv('Data/Raw data/weather.txt', sep=',')

#extract timestamp and 3 specific houses consumption for source, target and target_2

pd.set_option('display.max_rows', 20000)


#extract timestamp and 3 specific houses consumption for source, target and target_2

columns_to_keep = ['timestamp','Panther_lodging_Anastasia']
electricity_data_source = electricity_data[columns_to_keep]

columns_to_keep = ['timestamp','Peacock_lodging_Lou']
electricity_data_target = electricity_data[columns_to_keep]

columns_to_keep = ['timestamp','Robin_lodging_Dorthy']
electricity_data_target_2 = electricity_data[columns_to_keep]

#extract the metadata for the 3 specific houses and multiply the rows of the dataframes to match the rest

metadata_data_source = metadata_data[metadata_data['building_id'] == 'Panther_lodging_Anastasia']
metadata_data_target = metadata_data[metadata_data['building_id'] == 'Peacock_lodging_Lou']
metadata_data_target_2 = metadata_data[metadata_data['building_id'] == 'Robin_lodging_Dorthy']

columns_to_keep = ['sqm', 'lat', 'lng', 'yearbuilt']

metadata_data_source = metadata_data_source[columns_to_keep]
metadata_data_target = metadata_data_target[columns_to_keep]
metadata_data_target_2 = metadata_data_target_2[columns_to_keep]

new_row_source = pd.DataFrame([[2272.0, 28.517689, -81.379039, 1968.0]], columns=metadata_data_source.columns)
new_row_target = pd.DataFrame([[11895.7, 40.35, -74.7, 2000.0]], columns=metadata_data_target.columns)
new_row_target_2 = pd.DataFrame([[5842.9, 51.51879, -0.134556, 1960.0]], columns=metadata_data_target.columns)

for i in range(17543):
    metadata_data_source = pd.concat([metadata_data_source, new_row_source], ignore_index=True)
    metadata_data_target = pd.concat([metadata_data_target, new_row_target], ignore_index=True)
    metadata_data_target_2 = pd.concat([metadata_data_target_2, new_row_target_2], ignore_index=True)

# extract weather data for the regions of the 3 specific houses

weather_data_source = weather_data[weather_data['site_id'] == 'Panther']

weather_data_target = weather_data[weather_data['site_id'] == 'Cockatoo']
weather_data_target = weather_data_target.reset_index(drop=True)

weather_data_target_2 = weather_data[weather_data['site_id'] == 'Robin']
weather_data_target_2 = weather_data_target_2.reset_index(drop=True)

columns_to_keep = ['airTemperature','cloudCoverage','dewTemperature','windDirection','windSpeed']

weather_data_source = weather_data_source[columns_to_keep]
weather_data_target = weather_data_target[columns_to_keep]
weather_data_target_2 = weather_data_target_2[columns_to_keep]

#merge the 3 dataframes (EC, metadata, weather data) into 1 for each domain (source, target, target_2)

data_source = pd.concat([electricity_data_source, metadata_data_source, weather_data_source], axis = 1)
data_target = pd.concat([electricity_data_target, metadata_data_target, weather_data_target], axis = 1)
data_target_2 = pd.concat([electricity_data_target_2, metadata_data_target_2, weather_data_target_2], axis = 1)

#rename the building id column to 'Electrical_Consumption' relocate it as the last column

last_column = data_source.pop('Panther_lodging_Anastasia')
data_source.insert(10, 'Electrical_Consumption', last_column)

last_column = data_target.pop('Peacock_lodging_Lou')
data_target.insert(10, 'Electrical_Consumption', last_column)

last_column = data_target_2.pop('Robin_lodging_Dorthy')
data_target_2.insert(10, 'Electrical_Consumption', last_column)

data_source.to_pickle('Data/Final data/data_source.pkl')
data_target.to_pickle('Data/Final data/data_target.pkl')
data_target_2.to_pickle('Data/Final data/data_target_2.pkl')
