==> Notebooks/5) Model_Results_CNN-LSTM-Hybrid.ipynb <==
from sklearn.preprocessing import MinMaxScaler
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import joblib
import pickle
import json
import seaborn as sns



def base_results_graph(predictions_base, dataset, domain):

        if domain == 'source':
            true_measurements_unscaled = (scaler_source.inverse_transform(y_test_source))*(sqm_source/100)
        elif domain == 'target':
            true_measurements_unscaled = (scaler_target.inverse_transform(y_test_target))*(sqm_target/100)
        elif domain == 'target_2':
            true_measurements_unscaled = (scaler_target_2.inverse_transform(y_test_target_2))*(sqm_target_2/100)
        
        periods = [24, 7*24, 30*24]
        time_periods = ['Day','Week','Month']
        

        for period,time_period in zip(periods,time_periods):

            # Create a time index for the next 24 hours
            time_index = range(period)

            plt.figure(figsize=(20, 9))

            # Plot predicted values
            plt.plot(time_index, predictions_base[0:period], label='Predicted_base_{}'.format(domain), color = 'blue')

            # Plot true measurements
            plt.plot(time_index, true_measurements_unscaled[0:period], label='True', color = 'orange', linestyle='--', alpha = 0.6)

            # Add labels and title
            plt.xlabel('Time (hours)')
            plt.ylabel('Measurement')
            plt.title('CNN-LSTM-Hybrid Base model next {} predictions / true measurements in {} domain'.format(time_period, domain))
            plt.legend()

            # Save plot
            plt.savefig('Results/CNN-LSTM-Hybrid/{}/Next {} Results/Prediction Results/BASE_{}_results.png'.format(dataset,time_period,domain), dpi=500)

            # Show plot
            plt.show()



def error_ranges(predictions_base, predictions_TL, predictions_TL2, predictions_TL3, predictions_TL4, predictions_TL5, dataset, domain):

    if domain == 'target':
        true_measurements_unscaled = (scaler_target.inverse_transform(y_test_target))*(sqm_target/100)
    elif domain == 'target_2':
        true_measurements_unscaled = (scaler_target_2.inverse_transform(y_test_target_2))*(sqm_target_2/100)


    periods = [24, 7*24, 30*24]
    time_periods = ['Day','Week','Month']
        

    for period,time_period in zip(periods,time_periods):

        base_errors = np.abs(predictions_base[0:period] - true_measurements_unscaled[0:period])
        base_errors = base_errors.ravel()

        TL_errors = np.abs(predictions_TL[0:period] - true_measurements_unscaled[0:period])
        TL_errors = TL_errors.ravel()

        TL2_errors = np.abs(predictions_TL2[0:period] - true_measurements_unscaled[0:period])
        TL2_errors = TL2_errors.ravel()

        TL3_errors = np.abs(predictions_TL3[0:period] - true_measurements_unscaled[0:period])
        TL3_errors = TL3_errors.ravel()

        TL4_errors = np.abs(predictions_TL4[0:period] - true_measurements_unscaled[0:period])
        TL4_errors = TL4_errors.ravel()

        TL5_errors = np.abs(predictions_TL5[0:period] - true_measurements_unscaled[0:period])
        TL5_errors = TL5_errors.ravel()

        error_data = {
            'Without TL': base_errors,
            'TL Strategy 1': TL_errors,
            'TL Strategy 2': TL2_errors,
            'TL Strategy 3': TL3_errors,
            'TL Strategy 4': TL4_errors,
            'TL Strategy 5': TL5_errors
        }

        # Create the boxplot without fill colors
        plt.figure(figsize=(15, 9))
        boxplot = sns.boxplot(data=list(error_data.values()), showfliers=False, widths=0.3, color='whitesmoke')

        # Customize x-ticks
        plt.xticks(ticks=np.arange(len(error_data.keys())), labels=error_data.keys())
        

        medians = []
        medians.append(boxplot.lines[4])
        medians.append(boxplot.lines[9])
        medians.append(boxplot.lines[14])
        medians.append(boxplot.lines[19])
        medians.append(boxplot.lines[24])
        medians.append(boxplot.lines[29])

        # Customize median line to be orange and thicker
        for median in medians:  # The first 4 lines are whiskers, the next 4 are medians
            median.set_color('gold')
            median.set_linewidth(2.3)  # Thicker median line

        # Add title and labels
        plt.title("CNN-LSTM-Hybrid models error spread for next {} predictions in {} {} domain".format(time_period,dataset,domain), fontsize=14)
        plt.ylabel("error")
        
        plt.grid()

        # Save plot
        plt.savefig('Results/CNN-LSTM-Hybrid/{}/Next {} Results/Error Range Results/Error_ranges_{}.png'.format(dataset,time_period,domain), dpi=500)

        # Show the plot
        plt.show()




def comparative_results_graph(predictions_base, predictions_TL, predictions_TL2, predictions_TL3, predictions_TL4, predictions_TL5, dataset, domain):

        if domain == 'target':

            true_measurements_unscaled = (scaler_target.inverse_transform(y_test_target))*(sqm_target/100)
        
        elif domain == 'target_2':

            true_measurements_unscaled = (scaler_target_2.inverse_transform(y_test_target_2))*(sqm_target_2/100)


        periods = [24, 7*24, 30*24]
        time_periods = ['Day','Week','Month']    

        for period,time_period in zip(periods,time_periods):

            # Create a time index for the next 24 hours
            time_index = range(period)
            plt.figure(figsize=(20, 9))

            # Plot predicted values
            plt.plot(time_index, predictions_base[0:period], label='Predicted_BASE_{}'.format(domain),color = 'black', linewidth=0.9)
            plt.plot(time_index, predictions_TL[0:period], label='Predicted_TL1_{}'.format(domain),color = 'red', linewidth=0.9)
            plt.plot(time_index, predictions_TL2[0:period], label='Predicted_TL2_{}'.format(domain),color = 'green', linewidth=0.9)
            plt.plot(time_index, predictions_TL3[0:period], label='Predicted_TL3_{}'.format(domain),color = 'blue', linewidth=0.9)
            plt.plot(time_index, predictions_TL4[0:period], label='Predicted_TL4_{}'.format(domain),color = 'purple', linewidth=0.9)
            plt.plot(time_index, predictions_TL5[0:period], label='Predicted_TL5_{}'.format(domain),color = 'pink', linewidth=0.9)

            # Plot true measurements
            plt.plot(time_index, true_measurements_unscaled[0:period], label='True', color = 'orange', linestyle='--', alpha=0.6)

            # Add labels and title
            plt.xlabel('Time (hours)')
            plt.ylabel('Measurement')
            plt.title('CNN-LSTM-Hybrid models next {} predictions / true measurements in {} {} domain'.format(time_period,dataset,domain))
            plt.legend()

            # Save plot
            plt.savefig('Results/CNN-LSTM-Hybrid/{}/Next {} Results/Prediction Results/TL_BASE_{}_values_comparative_results.png'.format(dataset,time_period,domain), dpi=500)

            # Show plot
            plt.show()


def comparative_metrics_results(metric, dataset):

    #the metrics file has 12 elements. 0-4 are the metrics for the next day, 4-8 are the metrics for the next week and 8-12 are the metrics for the next month

    range = [0,4,8]
    time_periods = ['Day','Week','Month'] 

    for i,time_period in zip(range,time_periods):

        if dataset == '12M':

            RMSE_values = [metrics_base_source[i+0],metrics_base_target[i+0],metrics_base_target_2[i+0],metrics_TL1_target[i+0],metrics_TL1_target_2[i+0],metrics_TL2_target[i+0],metrics_TL2_target_2[i+0],metrics_TL3_target[i+0],metrics_TL3_target_2[i+0],metrics_TL4_target[i+0],metrics_TL4_target_2[i+0],metrics_TL5_target[i+0],metrics_TL5_target_2[i+0]]
            MAE_values = [metrics_base_source[i+1],metrics_base_target[i+1],metrics_base_target_2[i+1],metrics_TL1_target[i+1],metrics_TL1_target_2[i+1],metrics_TL2_target[i+1],metrics_TL2_target_2[i+1],metrics_TL3_target[i+1],metrics_TL3_target_2[i+1],metrics_TL4_target[i+1],metrics_TL4_target_2[i+1],metrics_TL5_target[i+1],metrics_TL5_target_2[i+1]]
            MAPE_values = [metrics_base_source[i+2],metrics_base_target[i+2],metrics_base_target_2[i+2],metrics_TL1_target[i+2],metrics_TL1_target_2[i+2],metrics_TL2_target[i+2],metrics_TL2_target_2[i+2],metrics_TL3_target[i+2],metrics_TL3_target_2[i+2],metrics_TL4_target[i+2],metrics_TL4_target_2[i+2],metrics_TL5_target[i+2],metrics_TL5_target_2[i+2]]
            R2_score_values = [metrics_base_source[i+3],metrics_base_target[i+3],metrics_base_target_2[i+3],metrics_TL1_target[i+3],metrics_TL1_target_2[i+3],metrics_TL2_target[i+3],metrics_TL2_target_2[i+3],metrics_TL3_target[i+3],metrics_TL3_target_2[i+3],metrics_TL4_target[i+3],metrics_TL4_target_2[i+3],metrics_TL5_target[i+3],metrics_TL5_target_2[i+3]]

            model_names = ['CNN-LSTM_base_source', 'CNN-LSTM_base_target', 'CNN-LSTM_base_target_2', 'CNN-LSTM_TL1_target', 'CNN-LSTM_TL1_target_2', 'CNN-LSTM_TL2_target', 'CNN-LSTM_TL2_target_2', 'CNN-LSTM_TL3_target', 'CNN-LSTM_TL3_target_2', 'CNN-LSTM_TL4_target', 'CNN-LSTM_TL4_target_2', 'CNN-LSTM_TL5_target', 'CNN-LSTM_TL5_target_2']

        if dataset == '6M':

            RMSE_values = [metrics_base_source[i+0],metrics_base_target[i+0],metrics_TL1_target[i+0],metrics_TL2_target[i+0],metrics_TL3_target[i+0],metrics_TL4_target[i+0],metrics_TL5_target[i+0]]
            MAE_values = [metrics_base_source[i+1],metrics_base_target[i+1],metrics_TL1_target[i+1],metrics_TL2_target[i+1],metrics_TL3_target[i+1],metrics_TL4_target[i+1],metrics_TL5_target[i+1]]
            MAPE_values = [metrics_base_source[i+2],metrics_base_target[i+2],metrics_TL1_target[i+2],metrics_TL2_target[i+2],metrics_TL3_target[i+2],metrics_TL4_target[i+2],metrics_TL5_target[i+2]]
            R2_score_values = [metrics_base_source[i+3],metrics_base_target[i+3],metrics_TL1_target[i+3],metrics_TL2_target[i+3],metrics_TL3_target[i+3],metrics_TL4_target[i+3],metrics_TL5_target[i+3]]

            model_names = ['CNN-LSTM_base_source', 'CNN-LSTM_base_target', 'CNN-LSTM_TL1_target', 'CNN-LSTM_TL2_target', 'CNN-LSTM_TL3_target', 'CNN-LSTM_TL4_target', 'CNN-LSTM_TL5_target']

        if metric == 'RMSE':
            plot_values = RMSE_values
        if metric == 'MAE':
            plot_values = MAE_values
        if metric == 'MAPE':
            plot_values = MAPE_values
        if metric == 'R2_Score':
            plot_values = R2_score_values

        plt.figure(figsize=(15, 9))
        plt.bar(model_names, plot_values)
        plt.xlabel('Model')
        plt.ylabel(metric)
        plt.title('{} values Comparison of CNN-LSTM-Hybrid models next {} predictions in {} datasets'.format(metric,time_period,dataset))
        plt.xticks(rotation=45, ha='right')  # Rotate x-axis labels and align them to the right
        plt.tight_layout()  # Adjust layout to make room for rotated labels

        # Save plot
        plt.savefig('Results/CNN-LSTM-Hybrid/{}/Next {} Results/Metrics Results/{}_values_comparative_results.png'.format(dataset,time_period,metric), dpi=500)

        plt.show()



def metrics_table(dataset):


    range = [0,4,8]
    time_periods = ['Day','Week','Month'] 

    for i,time_period in zip(range,time_periods):

        if dataset == '12M':

            RMSE_values = [metrics_base_source[i+0],metrics_base_target[i+0],metrics_base_target_2[i+0],metrics_TL1_target[i+0],metrics_TL1_target_2[i+0],metrics_TL2_target[i+0],metrics_TL2_target_2[i+0],metrics_TL3_target[i+0],metrics_TL3_target_2[i+0],metrics_TL4_target[i+0],metrics_TL4_target_2[i+0],metrics_TL5_target[i+0],metrics_TL5_target_2[i+0]]
            MAE_values = [metrics_base_source[i+1],metrics_base_target[i+1],metrics_base_target_2[i+1],metrics_TL1_target[i+1],metrics_TL1_target_2[i+1],metrics_TL2_target[i+1],metrics_TL2_target_2[i+1],metrics_TL3_target[i+1],metrics_TL3_target_2[i+1],metrics_TL4_target[i+1],metrics_TL4_target_2[i+1],metrics_TL5_target[i+1],metrics_TL5_target_2[i+1]]
            MAPE_values = [metrics_base_source[i+2],metrics_base_target[i+2],metrics_base_target_2[i+2],metrics_TL1_target[i+2],metrics_TL1_target_2[i+2],metrics_TL2_target[i+2],metrics_TL2_target_2[i+2],metrics_TL3_target[i+2],metrics_TL3_target_2[i+2],metrics_TL4_target[i+2],metrics_TL4_target_2[i+2],metrics_TL5_target[i+2],metrics_TL5_target_2[i+2]]
            R2_score_values = [metrics_base_source[i+3],metrics_base_target[i+3],metrics_base_target_2[i+3],metrics_TL1_target[i+3],metrics_TL1_target_2[i+3],metrics_TL2_target[i+3],metrics_TL2_target_2[i+3],metrics_TL3_target[i+3],metrics_TL3_target_2[i+3],metrics_TL4_target[i+3],metrics_TL4_target_2[i+3],metrics_TL5_target[i+3],metrics_TL5_target_2[i+3]]

            model_names = ['CNN-LSTM_base_source', 'CNN-LSTM_base_target', 'CNN-LSTM_base_target_2', 'CNN-LSTM_TL1_target', 'CNN-LSTM_TL1_target_2', 'CNN-LSTM_TL2_target', 'CNN-LSTM_TL2_target_2', 'CNN-LSTM_TL3_target', 'CNN-LSTM_TL3_target_2', 'CNN-LSTM_TL4_target', 'CNN-LSTM_TL4_target_2', 'CNN-LSTM_TL5_target', 'CNN-LSTM_TL5_target_2']

            metrics_df = pd.DataFrame({
            'Model': model_names,
            'RMSE': [RMSE_values[0], RMSE_values[1], RMSE_values[2], RMSE_values[3], RMSE_values[4], RMSE_values[5], RMSE_values[6], RMSE_values[7], RMSE_values[8], RMSE_values[9], RMSE_values[10], RMSE_values[11], RMSE_values[12]],
            'MAE': [MAE_values[0], MAE_values[1], MAE_values[2], MAE_values[3], MAE_values[4], MAE_values[5], MAE_values[6], MAE_values[7], MAE_values[8], MAE_values[9], MAE_values[10], MAE_values[11], MAE_values[12]],
            'MAPE': [MAPE_values[0], MAPE_values[1], MAPE_values[2], MAPE_values[3], MAPE_values[4], MAPE_values[5], MAPE_values[6], MAPE_values[7], MAPE_values[8], MAPE_values[9], MAPE_values[10], MAPE_values[11], MAPE_values[12]],
            'R2': [R2_score_values [0], R2_score_values [1], R2_score_values [2], R2_score_values [3], R2_score_values [4], R2_score_values [5], R2_score_values [6], R2_score_values [7], R2_score_values [8], R2_score_values [9], R2_score_values [10], R2_score_values [11], R2_score_values [12]]
            })

        elif dataset == '6M':

            RMSE_values = [metrics_base_source[i+0],metrics_base_target[i+0],metrics_TL1_target[i+0],metrics_TL2_target[i+0],metrics_TL3_target[i+0],metrics_TL4_target[i+0],metrics_TL5_target[i+0]]
            MAE_values = [metrics_base_source[i+1],metrics_base_target[i+1],metrics_TL1_target[i+1],metrics_TL2_target[i+1],metrics_TL3_target[i+1],metrics_TL4_target[i+1],metrics_TL5_target[i+1]]
            MAPE_values = [metrics_base_source[i+2],metrics_base_target[i+2],metrics_TL1_target[i+2],metrics_TL2_target[i+2],metrics_TL3_target[i+2],metrics_TL4_target[i+2],metrics_TL5_target[i+2]]
            R2_score_values = [metrics_base_source[i+3],metrics_base_target[i+3],metrics_TL1_target[i+3],metrics_TL2_target[i+3],metrics_TL3_target[i+3],metrics_TL4_target[i+3],metrics_TL5_target[i+3]]

            model_names = ['CNN-LSTM_base_source', 'CNN-LSTM_base_target', 'CNN-LSTM_TL1_target', 'CNN-LSTM_TL2_target', 'CNN-LSTM_TL3_target', 'CNN-LSTM_TL4_target', 'CNN-LSTM_TL5_target']

            metrics_df = pd.DataFrame({
            'Model': model_names,
            'RMSE': [RMSE_values[0], RMSE_values[1], RMSE_values[2], RMSE_values[3], RMSE_values[4], RMSE_values[5], RMSE_values[6]],
            'MAE': [MAE_values[0], MAE_values[1], MAE_values[2], MAE_values[3], MAE_values[4], MAE_values[5], MAE_values[6]],
            'MAPE': [MAPE_values[0], MAPE_values[1], MAPE_values[2], MAPE_values[3], MAPE_values[4], MAPE_values[5], MAPE_values[6]],
            'R2': [R2_score_values [0], R2_score_values [1], R2_score_values [2], R2_score_values [3], R2_score_values [4], R2_score_values [5], R2_score_values [6]]
            })

        metrics_df = metrics_df.round(3)

        # Plot the table
        fig, ax = plt.subplots(figsize=(15, 9))  # Set the size as per your requirement
        ax.axis('tight')
        ax.axis('off')
        table = ax.table(cellText=metrics_df.values, colLabels=metrics_df.columns, cellLoc='center', loc='center')

        # Adjust vertical space by setting row heights
        table.auto_set_font_size(False)
        table.set_fontsize(12)
        table.scale(1, 2)  # Scale width by 1, height by 2 (change 2 to a larger value for more height)

        for key, cell in table.get_celld().items():
            cell.set_fontsize(12)  # Adjust font size if needed
            cell.set_linewidth(0.5)  # Adjust line width if needed

        plt.tight_layout()
        if dataset == '12M':
            plt.suptitle('CNN-LSTM-Hybrid models metrics for next {} predictions in {} datasets'.format(time_period,dataset), y=0.85)
        elif dataset == '6M':
            plt.suptitle('CNN-LSTM-Hybrid models metrics for next {} predictions in {} datasets'.format(time_period,dataset), y=0.72)
        plt.savefig('Results/CNN-LSTM-Hybrid/{}/Next {} Results/Metrics Results/metrics_table.png'.format(dataset,time_period), bbox_inches='tight', dpi=500)
        plt.show()

#Results for 12M datasets


scaler_source = joblib.load('Scalers/12M/source_scaler.pkl')
scaler_target = joblib.load('Scalers/12M/target_scaler.pkl')
scaler_target_2 = joblib.load('Scalers/12M/target_2_scaler.pkl')

with open('Domain sqms/sqm_source.json', 'r') as f:
    sqm_source = json.load(f)

with open('Domain sqms/sqm_target.json', 'r') as f:
    sqm_target = json.load(f)

with open('Domain sqms/sqm_target_2.json', 'r') as f:
    sqm_target_2 = json.load(f)


y_test_source = np.load('Data/Preprocessed data/12M/source_test_y.npy')
y_test_source = y_test_source[:-24]
y_test_target = np.load('Data/Preprocessed data/12M/target_test_y.npy')
y_test_target = y_test_target[:-24]
y_test_target_2 = np.load('Data/Preprocessed data/12M/target_2_test_y.npy')
y_test_target_2 = y_test_target_2[:-24]

predictions_base_source = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_base_source.npy')
predictions_base_target = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_base_target.npy')
predictions_base_target_2 = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_base_target_2.npy')
predictions_TL1_target = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL1_target.npy')
predictions_TL1_target_2 = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL1_target_2.npy')
predictions_TL2_target = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL2_target.npy')
predictions_TL2_target_2 = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL2_target_2.npy')
predictions_TL3_target = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL3_target.npy')
predictions_TL3_target_2 = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL3_target_2.npy')
predictions_TL4_target = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL4_target.npy')
predictions_TL4_target_2 = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL4_target_2.npy')
predictions_TL5_target = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL5_target.npy')
predictions_TL5_target_2 = np.load('Models/CNN-LSTM-Hybrid/12M/Predictions/predictions_TL5_target_2.npy')

with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_base_source.pkl', 'rb') as file:
    metrics_base_source = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_base_target.pkl', 'rb') as file:
    metrics_base_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_base_target_2.pkl', 'rb') as file:
    metrics_base_target_2 = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL1_target.pkl', 'rb') as file:
    metrics_TL1_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL1_target_2.pkl', 'rb') as file:
    metrics_TL1_target_2 = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL2_target.pkl', 'rb') as file:
    metrics_TL2_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL2_target_2.pkl', 'rb') as file:
    metrics_TL2_target_2 = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL3_target.pkl', 'rb') as file:
    metrics_TL3_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL3_target_2.pkl', 'rb') as file:
    metrics_TL3_target_2 = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL4_target.pkl', 'rb') as file:
    metrics_TL4_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL4_target_2.pkl', 'rb') as file:
    metrics_TL4_target_2 = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL5_target.pkl', 'rb') as file:
    metrics_TL5_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/12M/Metrics/metrics_TL5_target_2.pkl', 'rb') as file:
    metrics_TL5_target_2 = pickle.load(file)


comparative_metrics_results('RMSE', '12M')

comparative_metrics_results('MAE', '12M')

comparative_metrics_results('MAPE', '12M')

comparative_metrics_results('R2_Score', '12M')

error_ranges(predictions_base_source, predictions_TL1_target, predictions_TL2_target, predictions_TL3_target, predictions_TL4_target, predictions_TL5_target, '12M', 'target')

error_ranges(predictions_base_target_2, predictions_TL1_target_2, predictions_TL2_target_2, predictions_TL3_target_2, predictions_TL4_target_2, predictions_TL5_target_2, '12M', 'target_2')

metrics_table('12M')

base_results_graph(predictions_base_source,'12M', 'source')

base_results_graph(predictions_base_target,'12M', 'target')

base_results_graph(predictions_base_target_2,'12M', 'target_2')

comparative_results_graph(predictions_base_target, predictions_TL1_target, predictions_TL2_target, predictions_TL3_target, predictions_TL4_target, predictions_TL5_target, '12M', 'target')

comparative_results_graph(predictions_base_target_2, predictions_TL1_target_2, predictions_TL2_target_2, predictions_TL3_target_2, predictions_TL4_target_2, predictions_TL5_target_2, '12M', 'target_2')


#Results for 6M datasets


scaler_target = joblib.load('Scalers/6M/target_scaler.pkl')


with open('Domain sqms/sqm_target.json', 'r') as f:
    sqm_target = json.load(f)


y_test_target = np.load('Data/Preprocessed data/6M/target_test_y.npy')
y_test_target = y_test_target[:-24]


predictions_base_target = np.load('Models/CNN-LSTM-Hybrid/6M/Predictions/predictions_base_target.npy')
predictions_TL1_target = np.load('Models/CNN-LSTM-Hybrid/6M/Predictions/predictions_TL1_target.npy')
predictions_TL2_target = np.load('Models/CNN-LSTM-Hybrid/6M/Predictions/predictions_TL2_target.npy')
predictions_TL3_target = np.load('Models/CNN-LSTM-Hybrid/6M/Predictions/predictions_TL3_target.npy')
predictions_TL4_target = np.load('Models/CNN-LSTM-Hybrid/6M/Predictions/predictions_TL4_target.npy')
predictions_TL5_target = np.load('Models/CNN-LSTM-Hybrid/6M/Predictions/predictions_TL5_target.npy')


with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_base_source.pkl', 'rb') as file:
    metrics_base_source = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_base_target.pkl', 'rb') as file:
    metrics_base_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_TL1_target.pkl', 'rb') as file:
    metrics_TL1_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_TL2_target.pkl', 'rb') as file:
    metrics_TL2_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_TL3_target.pkl', 'rb') as file:
    metrics_TL3_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_TL4_target.pkl', 'rb') as file:
    metrics_TL4_target = pickle.load(file)
with open('Models/CNN-LSTM-Hybrid/6M/Metrics/metrics_TL5_target.pkl', 'rb') as file:
    metrics_TL5_target = pickle.load(file)


comparative_metrics_results('RMSE', '6M')

comparative_metrics_results('MAE', '6M')

comparative_metrics_results('MAPE', '6M')

comparative_metrics_results('R2_Score', '6M')

error_ranges(predictions_base_source, predictions_TL1_target, predictions_TL2_target, predictions_TL3_target, predictions_TL4_target, predictions_TL5_target, '6M', 'target')

metrics_table('6M')

base_results_graph(predictions_base_target, '6M', 'target')

comparative_results_graph(predictions_base_target, predictions_TL1_target, predictions_TL2_target, predictions_TL3_target, predictions_TL4_target, predictions_TL5_target, '6M', 'target')
